Swap order of rabbitmqadmin tests
On Ubuntu, the rabbitmq-server package will automatically start the
service, but the 'does nothing if service is unmanaged' will change the
rabbitmq config file without restarting the service. The following
test, 'downloads the cli tools' will also not refresh the service
because the config file has not changed since the last test, but the
server will not be listening on the right port and the staging resource
fill fail. This patch swaps the order of the tests so that the service
will be listening correctly on the first test, but then will correctly
do nothing on the second test.